requirements.txt
# CourseCompanion - Python Dependencies
# Install with: pip install -r requirements.txt

# ==================== Core Framework ====================
fastapi==0.109.0
uvicorn[standard]==0.27.0
pydantic==2.5.3
pydantic-settings==2.1.0

# ==================== Frontend ====================
streamlit==1.30.0
streamlit-extras==0.3.6

# ==================== Database ====================
motor==3.3.2
pymongo==4.6.1

# ==================== AI/ML - LangChain ====================
langchain==0.1.4
langchain-openai==0.0.5
langchain-community==0.0.16
langchain-mongodb==0.1.0
langgraph==0.0.20

# ==================== OpenAI ====================
openai==1.10.0
tiktoken==0.5.2

# ==================== HTTP & Async ====================
httpx==0.26.0
aiohttp==3.9.3
requests==2.31.0

# ==================== Utilities ====================
python-dotenv==1.0.1
python-multipart==0.0.6
tenacity==8.2.3

# ==================== Data Processing ====================
numpy==1.26.3
pandas==2.2.0

# ==================== Development ====================
pytest==7.4.4
pytest-asyncio==0.23.4
black==24.1.1
isort==5.13.2
flake8==7.0.0

# ==================== Optional - Vector Search ====================
# Uncomment if using local vector search instead of MongoDB Atlas
# faiss-cpu==1.7.4
# chromadb==0.4.22

